NetworkDesign.jpeg
I chose an infrastructure topology. I chose this because I wanted to make something that used a combination of wired and wireless devices. Wired devices in my lucidchart are connected to a switch. A wireless access point is then connected to a laptop, wirelessly. The access point here works as almost a bridge between the wireless and wired networks. I could have put in more access points but I chose to make it the way I did to just show a basic infrastructure topology.